REFACTOR: re-package for ncp object
